Правки к dom4. создала три функции, изменила index
