+++ index.html
@@ -12,8 +12,7 @@
     <ul class="comments">
       <li class="comment">
         <div class="comment-header">
-          <div>Глеб Фокин</div>
-          <div>12.02.22 12:18</div>
+           <div>12.02.22 12:18</div>
         </div>
         <div class="comment-body">
           <div class="comment-text">
