+++ module/handleAddComment.js
@@ -1,0 +1,41 @@
+export function handleAddComment(commentsData, renderComments) {
+    const addButton = document.querySelector(".add-form-button");
+    const nameInput = document.querySelector(".add-form-name");
+    const commentTextarea = document.querySelector(".add-form-text");
+  
+    if (!addButton || !nameInput || !commentTextarea) return;
+  
+    addButton.addEventListener("click", (e) => {
+      e.preventDefault();
+      const name = nameInput.value.trim();
+      const text = commentTextarea.value.trim();
+  
+      if (!name || !text) {
+        alert("Заполните поля!");
+        return;
+      }
+  
+      const now = new Date();
+      const day = String(now.getDate()).padStart(2, "0");
+      const month = String(now.getMonth() + 1).padStart(2, "0");
+      const year = String(now.getFullYear()).slice(-2);
+      const hours = String(now.getHours()).padStart(2, "0");
+      const minutes = String(now.getMinutes()).padStart(2, "0");
+      const formattedDate = `${day}.${month}.${year} ${hours}:${minutes}`;
+  
+      const newComment = {
+        id: Date.now(),
+        name: name,
+        date: formattedDate,
+        text: text,
+        likesCount: 0,
+        isLiked: false,
+      };
+  
+      commentsData.push(newComment);
+      renderComments(commentsData);
+  
+      nameInput.value = "";
+      commentTextarea.value = "";
+    });
+  }
